Megre Tanvn into main

+++ laravel/app/Http/Controllers/Api/V1/User/UserCatalogueController.php
@@ -31,7 +31,7 @@ class UserCatalogueController extends Controller
      */
     public function index(): JsonResponse
     {
-        // $this->authorize('modules', 'users.catalogues.index');
+        $this->authorize('modules', 'users.catalogues.index');
 
         $paginator = $this->userCatalogueService->paginate();
         $data = new UserCatalogueCollection($paginator);
@@ -44,7 +44,7 @@ class UserCatalogueController extends Controller
      */
     public function store(StoreUserCatalogueRequest $request): JsonResponse
     {
-        // $this->authorize('modules', 'users.catalogues.store');
+        $this->authorize('modules', 'users.catalogues.store');
 
         $response = $this->userCatalogueService->create();
 
@@ -56,7 +56,7 @@ class UserCatalogueController extends Controller
      */
     public function show(string $id): JsonResponse
     {
-        // $this->authorize('modules', 'users.catalogues.show');
+        $this->authorize('modules', 'users.catalogues.show');
 
         $response = new UserCatalogueResource($this->userCatalogueRepository->findById($id));
 
@@ -68,7 +68,7 @@ class UserCatalogueController extends Controller
      */
     public function update(UpdateUserCatalogueRequest $request, string $id): JsonResponse
     {
-        // $this->authorize('modules', 'users.catalogues.update');
+        $this->authorize('modules', 'users.catalogues.update');
 
         $response = $this->userCatalogueService->update($id);
 
@@ -80,7 +80,7 @@ class UserCatalogueController extends Controller
      */
     public function destroy(string $id): JsonResponse
     {
-        // $this->authorize('modules', 'users.catalogues.destroy');
+        $this->authorize('modules', 'users.catalogues.destroy');
 
         $response = $this->userCatalogueService->destroy($id);
 
@@ -92,7 +92,7 @@ class UserCatalogueController extends Controller
      */
     public function updatePermissions(string $id): JsonResponse
     {
-        // $this->authorize('modules', 'users.catalogues.updatePermissions');
+        $this->authorize('modules', 'users.catalogues.updatePermissions');
 
         $response = $this->userCatalogueService->updatePermissions();
